refactor(settings): one canonical buildProfileMessage owner for both hosts
UPDATE_PROFILE was hand-assembled separately in the extension and the desktop,
with the UpdateProfileMessage type referenced nowhere — so the compiler checked
neither side, and they had drifted: desktop silently dropped spendingStatus and
quotaAutoSwitched, leaving the RelayQuotaMeter / quota-auto-switch banner dead on
desktop.

Introduce src/shared/settingsView/handlers/profileHandlers.buildProfileMessage()
(mirroring buildHistoryMessage): one host-neutral projection over auth / tier /
server-side-key / agent services returning the typed UpdateProfileMessage. The
only host-specific input — provider key statuses (VS Code settings on the
extension, none on desktop) — is injected as a callback. Both hosts now post
`await buildProfileMessage({ getProviderKeyStatuses })`, so the compiler enforces
all 13 fields on both sides and desktop gets spendingStatus + quotaAutoSwitched.

The builder keeps the desktop's graceful try/catch resilience (a transient
metadata failure keeps the user signed in with fallbacks) — the extension gains
that resilience — and primes the server-side-key cache via canUseServerSideKeys()
as both hosts already did.

Deletions: the extension sendProfileData inline assembly + primeIncludedAccessIfEnabled;
desktop loadDesktopAuthProfileData + unauthenticatedProfileData + DesktopAuthProfileData
+ DesktopSupabaseAuth.getProfileData + the getAuthProfileData IPC option. The two
index.ts getProfileData consumers (login banner + getAuthStatus) only read
.authenticated, so they move to SupabaseClient.isAuthenticated() directly.

Net -170 LOC. tsc (root + extension + desktop), prettier, eslint clean;
HistoryHandlers, DesktopSettingsIpc (33/33), DesktopSettingsSurface pass.
